v0.2-MCP-Tools/requirements.txt
mcp[cli]>=1.0.0
psycopg2-binary>=2.9.0